+++ app/services/init/__init__.py
@@ -4,8 +4,8 @@ Initialization Services Package
 Contains application initialization services.
 """
 
-from .startup_init import startup_init
-from .background_queue_init import background_queue_init
+from .startup_init import initialize_background_services as startup_init
+from .background_queue_init import initialize_background_queue as background_queue_init
 
 __all__ = [
     'startup_init',
