+++ app/services/recording/recording_logger.py
@@ -6,7 +6,7 @@ import uuid
 import json
 from typing import Optional, Any
 
-from app.services.logging_service import logging_service
+from app.services.system.logging_service import logging_service
 
 recording_logger = logging.getLogger("streamvault.recording")
